Created transaction result also for negative cases
