+++ app/src/main/java/network/o3/o3wallet/Wallet/SendV2/SendSuccessFragment.kt
@@ -121,16 +121,6 @@ class SendSuccessFragment : Fragment() {
         }
     }
 
-    fun setTransactionResult(sendViewModel: SendViewModel) {
-        val intent = Intent()
-        intent.putExtra("assetName", sendViewModel.getSelectedAsset().value!!.name)
-        intent.putExtra("amount", sendViewModel.getSelectedSendAmount().toDouble())
-        intent.putExtra("address", sendViewModel.getSelectedAddress().value!!)
-        intent.putExtra("transactionId", sendViewModel.txID)
-
-        activity?.setResult(Activity.RESULT_OK, intent)
-    }
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
 
@@ -141,7 +131,6 @@ class SendSuccessFragment : Fragment() {
         initiateSelectedRecipientDetails()
         initiateSelectedAssetDetails()
         initiateActionButtons()
-        setTransactionResult((activity as SendV2Activity).sendViewModel)
 
         return mView
     }
